requirements.txt
Flask
flask-cors
requests
python-dotenv
youtube_transcript_api
google-generativeai
